requirements.txt
moviepy==1.0.3
edge-tts
asyncio
numpy<2.0
Pillow<10.0.0
proglog
